Retire the next-reconcile job and dead calendar machinery
FO-authorized, captain-approved scope change: the edge marketplace entry now
resolves `main` directly instead of `next` (team-lead, commit e869a67 in the
standalone marketplace repo, live-probed: claude plugin marketplace update +
plugin update moved an installed plugin 0.27.0 -> 0.27.0-pre7, contract
byte-identical to main). `next` was only ever a second copy of `main` that a CI
job kept synced by comparing a tag's target version against next's manifest --
exactly the comparison that silently skipped v0.27.0-pre5 through -pre7 once a
bare 0.27.0 poisoned next. Deleting the comparison (and the reconcile it gated)
removes the entire silent-skip failure class instead of hardening a path that
no longer needs to exist.

release.yml's edge-advance job: both -X theirs reconcile steps and the
calendar-bump-and-push-to-next step are gone. The auto-pre0 cut (still real
value: catches the edge binary's minor up to a new stable line within minutes)
and the spacedock@next cask bump survive unchanged. The auto-pre0 step still
needs an "is this the latest known release line" check -- an old-line patch tag
would otherwise auto-cut a wrong, lower pre0 and bump the cask DOWN a minor --
so the decision step now compares the tag against the highest OTHER existing
bare stable git tag instead of next's manifest, reusing EdgeAdvanceDecision/
ComparePreVersion unchanged. Proposed to the FO before implementing; flagged in
the entity for review in case a different mechanism is preferred.

bump-calendar (the Go function, CLI command, and next-publish.yml, whose only
job was that bump) is deleted outright: the marketplace entry's version field
is inert on both hosts (round-2 measurement), so it was never a real re-pull
key. internal/release/edge_reconcile_test.go and edge_advance_noregress_test.go
are deleted wholesale -- the mechanism they proved no longer exists; two shared
helpers (edgeAdvanceJob, ifSelectsStable) relocate into edge_advance_wiring_test.go,
which is trimmed to the surviving pre0-only guard. docs/releasing.md, AGENTS.md
(CLAUDE.md is a symlink to it) rewritten to name main as the real re-pull key.

go test ./... and go test ./internal/release/... ./cmd/spacedock-release/... -race
both green (the one pre-existing internal/cli failure from the earlier commit
on this branch is unrelated and still held pending the captain's decision).
