--- .github/workflows/next-publish.yml
@@ -1,38 +1,0 @@
-# Bumps the marketplace entry's calendar version on a deliberate `next` publish.
-# The calendar key (`0.0.YYYYMMDDNN`) is the `claude plugin update` re-pull key:
-# bumping it is the ONLY thing that makes `update` re-pull the moving `next`
-# branch. Triggered manually (`workflow_dispatch`) — run it when you want `@next`
-# marketplace installers to re-pull, not on every push.
-name: next-publish
-
-on:
-  workflow_dispatch:
-
-permissions:
-  contents: write
-
-jobs:
-  bump-calendar:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v5
-
-      - name: Set up Go
-        uses: actions/setup-go@v6
-        with:
-          go-version: "1.22"
-
-      - name: Bump marketplace calendar version
-        run: |
-          set -euo pipefail
-          go run ./cmd/spacedock-release bump-calendar .claude-plugin/marketplace.json
-          if git diff --quiet -- .claude-plugin/marketplace.json; then
-            echo "calendar version unchanged; nothing to commit"
-            exit 0
-          fi
-          git config user.name "github-actions[bot]"
-          git config user.email "github-actions[bot]@users.noreply.github.com"
-          git commit -m "next: bump marketplace calendar version" \
-            -- .claude-plugin/marketplace.json
-          git push origin next
